backend/pyproject.toml
[tool.poetry]
name = "verspaetungsorakel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.16.2"
pymysql = "^1.1.0"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
xmltodict = "^0.13.0"
flask-cors = "^4.0.0"
pytest = "^7.4.0"
flask-limiter = "^3.3.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
